+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Reorganized preferences
 
+### Removed
+
+- FPS limit setting
+
 ## [0.5.0] - 2022-05-29
 
 ### Added